# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API documentation for the Buyverse Platform",
    "version": "1.0.0",
    "title": "Buyverse API",
    "contact": {
      "email": "support@buyverse.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/verify-otp": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Verify OTP",
        "description": "Endpoint to verify OTP received by the user.",
        "operationId": "verifyOTP",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "OTP verification data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyOTP"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products",
        "description": "Endpoint to retrieve all products.",
        "operationId": "getProducts",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "400": {
            "description": "Error occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get a single product",
        "description": "Endpoint to retrieve a single product by ID.",
        "operationId": "getProduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "404": {
            "description": "Product not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/products/category/{category}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get products by category",
        "description": "Endpoint to retrieve products by category.",
        "operationId": "getCategoryProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Product category"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products in the category",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "404": {
            "description": "No products found in this category",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/products/price/{min}/{max}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get products within a price range",
        "description": "Endpoint to retrieve products within a specific price range.",
        "operationId": "getPriceRangeProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "min",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Minimum price"
          },
          {
            "name": "max",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Maximum price"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products within the price range",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "404": {
            "description": "No products found within this price range",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/products/seller/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get products by seller",
        "description": "Endpoint to retrieve products by a specific seller.",
        "operationId": "getSellerProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Seller ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products by the seller",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "404": {
            "description": "Seller not found or no products found for this seller",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "VerifyOTP": {
      "type": "object",
      "required": [
        "email",
        "otp"
      ],
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "example": "john.doe@example.com"
        },
        "otp": {
          "type": "string",
          "example": "123456"
        }
      }
    },
...
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Operation completed successfully"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "An error occurred"
        },
        "error": {
          "type": "string",
          "example": "Detailed error message"
        }
      }
    },
...
    "Product": {
      "type": "object",
      "required": [
        "name",
        "price",
        "description",
        "category",
        "seller"
      ],
      "properties": {
        "id": {
          "type": "string",
          "example": "prod123"
        },
        "name": {
          "type": "string",
          "example": "Gadget Pro"
        },
        "price": {
          "type": "number",
          "format": "double",
          "example": 299.99
        },
        "description": {
          "type": "string",
          "example": "Latest model of the Gadget Pro with enhanced features."
        },
        "category": {
          "type": "string",
          "example": "Electronics"
        },
        "seller": {
          "type": "string",
          "example": "seller123"
        },
        "stock": {
          "type": "integer",
          "example": 150
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "http://example.com/image1.jpg"
          }
        }
      }
    },
...
  }
}